+++ src/pages/Teacher/TeacherCourses.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { ICourse } from '../../interfaces/components/entities/ICourse'
+import { ICourse } from '../../interfaces/components/entities'
 
 import { courseService } from '../../services'
 import ReusableTable from '../../Components/common/tables/ReusableTable'
